feat: 게임 종료 이벤트 핸들러 추가
